requirements.txt
nats-py
fastapi 
uvicorn 
websockets
fastapi-cors
